# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

/**
 * Whether diagnosing a crop or livestock
 */
export type DiagnoseInputSubjectType = typeof DiagnoseInputSubjectType[keyof typeof DiagnoseInputSubjectType];


export const DiagnoseInputSubjectType = {
  crop: 'crop',
  livestock: 'livestock',
} as const;

export interface DiagnoseInput {
  /** Base64 encoded image (optional) */
  imageBase64?: string;
  /** Text description of the problem */
  description: string;
  /** Whether diagnosing a crop or livestock */
  subjectType: DiagnoseInputSubjectType;
  /** Type of crop (e.g. maize, wheat, tomato) */
  cropType?: string;
  /** Type of livestock (e.g. cattle, chicken, pig) */
  livestockType?: string;
}

export type DiagnoseResultSeverity = typeof DiagnoseResultSeverity[keyof typeof DiagnoseResultSeverity];


export const DiagnoseResultSeverity = {
  low: 'low',
  medium: 'medium',
  high: 'high',
  critical: 'critical',
} as const;

export type DiagnoseResultConfidence = typeof DiagnoseResultConfidence[keyof typeof DiagnoseResultConfidence];


export const DiagnoseResultConfidence = {
  low: 'low',
  medium: 'medium',
  high: 'high',
} as const;

export interface DiagnoseResult {
  condition: string;
  severity: DiagnoseResultSeverity;
  confidence: DiagnoseResultConfidence;
  summary: string;
  symptoms: string[];
  causes: string[];
  treatments: string[];
  prevention: string[];
  urgency: string;
}

export interface ApiError {
  error: string;
}

export interface AnthropicConversation {
  id: number;
  title: string;
  createdAt: string;
}

export interface AnthropicMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface AnthropicConversationInput {
  title: string;
}

export interface AnthropicMessageInput {
  content: string;
}

export interface AnthropicConversationWithMessages {
  id: number;
  title: string;
  createdAt: string;
  messages: AnthropicMessage[];
}

export interface AnthropicError {
  error: string;
}