# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /diagnose:
    post:
      operationId: diagnose
      tags: [diagnose]
      summary: Diagnose crop or livestock issue
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DiagnoseInput"
      responses:
        "200":
          description: Diagnosis result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DiagnoseResult"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
components:
  schemas:
...
    DiagnoseInput:
      type: object
      properties:
        imageBase64:
          type: string
          description: Base64 encoded image (optional)
        description:
          type: string
          description: Text description of the problem
        subjectType:
          type: string
          enum: [crop, livestock]
          description: Whether diagnosing a crop or livestock
        cropType:
          type: string
          description: Type of crop (e.g. maize, wheat, tomato)
        livestockType:
          type: string
          description: Type of livestock (e.g. cattle, chicken, pig)
      required:
        - description
        - subjectType
    DiagnoseResult:
      type: object
      properties:
        condition:
          type: string
        severity:
          type: string
          enum: [low, medium, high, critical]
        confidence:
          type: string
          enum: [low, medium, high]
        summary:
          type: string
        symptoms:
          type: array
          items:
            type: string
        causes:
          type: array
          items:
            type: string
        treatments:
          type: array
          items:
            type: string
        prevention:
          type: array
          items:
            type: string
        urgency:
          type: string
      required:
        - condition
        - severity
        - confidence
        - summary
        - symptoms
        - causes
        - treatments
        - prevention
        - urgency
    ApiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...