# lib/api-zod/src/generated/types/diagnoseInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DiagnoseInputSubjectType } from './diagnoseInputSubjectType';

export interface DiagnoseInput {
  /** Base64 encoded image (optional) */
  imageBase64?: string;
  /** Text description of the problem */
  description: string;
  /** Whether diagnosing a crop or livestock */
  subjectType: DiagnoseInputSubjectType;
  /** Type of crop (e.g. maize, wheat, tomato) */
  cropType?: string;
  /** Type of livestock (e.g. cattle, chicken, pig) */
  livestockType?: string;
}