# lib/api-zod/src/generated/types/diagnoseResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DiagnoseResultConfidence } from './diagnoseResultConfidence';
import type { DiagnoseResultSeverity } from './diagnoseResultSeverity';

export interface DiagnoseResult {
  condition: string;
  severity: DiagnoseResultSeverity;
  confidence: DiagnoseResultConfidence;
  summary: string;
  symptoms: string[];
  causes: string[];
  treatments: string[];
  prevention: string[];
  urgency: string;
}